Revert to Control key for Xpad built-in window operations (Control-drag to move, Control-right clock drag to resize).
Trying to use Command for this obviously interferes with the window
manager, Option does not work for some reason, and Control does not
interfere with anything. We just need to be able to resize the stickies
somehow without relying on a Mac-style resize widget in the toolbar,
because the toolbar is not always visible.